+++ obj/test/services/DummyCommandableAzureFunctionService.test.js
@@ -17,7 +17,7 @@ suite('DummyCommandableAzureFunctionService', () => {
     let DUMMY2 = { id: null, key: "Key 2", content: "Content 2" };
     let _functionService;
     suiteSetup(() => __awaiter(void 0, void 0, void 0, function* () {
-        let config = pip_services3_commons_nodex_1.ConfigParams.fromTuples('logger.descriptor', 'pip-services:logger:console:default:1.0', 'controller.descriptor', 'pip-services-dummies:controller:default:default:1.0', 'service.descriptor', 'pip-services-dummies:service:commandable-azure-function:default:1.0');
+        let config = pip_services3_commons_nodex_1.ConfigParams.fromTuples('logger.descriptor', 'pip-services:logger:console:default:1.0', 'controller.descriptor', 'pip-services-dummies:controller:default:default:1.0', 'service.descriptor', 'pip-services-dummies:service:commandable-azurefunc:default:1.0');
         _functionService = new DummyCommandableAzureFunction_1.DummyCommandableAzureFunction();
         _functionService.configure(config);
         yield _functionService.open(null);