+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200">
 <br/> Azure specific components for Node.js / ES2017 Changelog
 
+## <a name="1.1.0"></a> 1.1.0 (2023-03-01)
+
+### Breaking changes
+* Renamed descriptors for services:
+    - "\*:service:azure-function\*:1.0" -> "\*:service:azurefunc\*:1.0"
+    - "\*:service:commandable-azure-function\*:1.0" -> "\*:service:commandable-azurefunc\*:1.0"
+
+### Features
+- Updated dependencies
+
 ## <a name="1.0.4"></a> 1.0.4 (2022-09-08)
 
 ### Bug fixes